% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionOutlier.R
\name{evolutionOutlier}
\alias{evolutionOutlier}
\title{Evolution outlier detection method}
\usage{
evolutionOutlier(dataMatrix, alpha)
}
\arguments{
\item{dataMatrix}{matrix with days by columns and time of the day by row or 3dimensional array with 3dim as smart meter indicator}

\item{alpha}{threhold for the outlier detection rule}
}
\value{
a list with the outlier index, distances to the prototype and others.
}
\description{
Compute the prototype, the distances to the prototype and provide the index of the detected outliers.
}
