% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnImpute.R
\name{knnImpute}
\alias{knnImpute}
\title{K-nearest neighbors imputation}
\usage{
knnImpute(dataMatrix, k = 1)
}
\arguments{
\item{dataMatrix}{functions by column and missing values as NAs}

\item{k}{number of neighbors}
}
\value{
matrix with imputed values
}
\description{
It uses similar days of the same meter to impute the missing data of a partially observe daily function
}
