% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnImputeFunction.R
\name{knnImputeFunction}
\alias{knnImputeFunction}
\title{K-nearest neighbors imputation}
\usage{
knnImputeFunction(data, k = 1)
}
\arguments{
\item{data}{3dimensional array}

\item{k}{number of neighbors. Default is 1.}
}
\value{
3dimensional array with imputed values
}
\description{
It uses similar meters of the same meter to impute the missing daily function
}
