% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveSmoothingBspline.R
\name{positiveSmoothingBspline}
\alias{positiveSmoothingBspline}
\title{Positive smoothing with Bsplines}
\usage{
positiveSmoothingBspline(matrixdata)
}
\arguments{
\item{matrixdata}{}
}
\value{
a list containing two matrices with the smooth functions and the first derivatives
}
\description{
Positive smoothing with Bsplines
}
\examples{

}
