% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessSolarData.R
\name{preProcessSolarData}
\alias{preProcessSolarData}
\title{Preprocess solar generation}
\usage{
preProcessSolarData(data, k)
}
\arguments{
\item{data}{matrix or 3dimensional array.}

\item{k}{if missing data, number of neighbors for the imputation method.}
}
\value{
a list containing three matrices with the raw curves, smooth functions and the first derivatives
}
\description{
Preprocess solar generation
}
