% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFarmData.R
\docType{data}
\name{solarPlant1}
\alias{solarPlant1}
\title{Solar Plant Data From Kaggle.}
\format{
A list with three 3d arrays. The list elements are 1) raw data, 2) level smooth data and 3) first derivarive of the smooth data.
}
\usage{
solarPlant1
}
\description{
Solar Plant Data From Kaggle.
}
\examples{
matplot(solarPlant1$raw, type = "l")
}
\references{
Kanal, A. (2020). "Kaggle:Solar powerg eneration data", \url{https://www.kaggle.com/anikannal/solar-power-generation-data/metadata}.
}
\keyword{datasets}
