% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDepth.R
\name{timeDepth}
\alias{timeDepth}
\title{Compute Fuctional Depths and Transformations}
\usage{
timeDepth(dataMatrix, typeDepth)
}
\arguments{
\item{dataMatrix}{matrix with days by columns and time of the day by row or 3dimensional array with 3dim as smart meter indicator}

\item{typeDepth}{depth measure to compute}
}
\description{
One meter data
}
